listados  de usuarios y volumen criptoactivos

+++ src/main/kotlin/ar/edu/unq/desapp/grupoL/criptop2p/persistence/UserRepository.kt
@@ -10,9 +10,9 @@ import java.util.*
 
 @Configuration
 @Repository
-interface UserRepository : CrudRepository<Usuario?, Int?> {
-    override fun findById(id: Int): Optional<Usuario?>
+interface UserRepository : CrudRepository<Usuario?, Long?> {
+    override fun findById(id: Long): Optional<Usuario?>
     fun save(usuario: Usuario): Usuario
      override fun findAll(): List<Usuario>
-     override fun deleteById(id: Int)
+     override fun deleteById(id: Long)
 }
